change msg type

+++ ros2_evcam_driver/include/ros2_evcam_driver/Visualization.hpp
@@ -5,7 +5,7 @@
 #include <cstdint>
 #include <opencv2/opencv.hpp>
 
-// bins対応済み visualizeHistogram（統一関数）
-cv::Mat visualizeHistogram(const std::vector<uint64_t>& hist, int bins, int width, int height, bool use_ratio = true);
+// bins対応済み visualizeHistogram（uint8_tベース）
+cv::Mat visualizeHistogram(const std::vector<uint8_t>& hist, int bins, int width, int height, bool use_ratio = true);
 
 #endif // VISUALIZATION_HPP
